add tests and fixes on tracepropagator handler
